Add previews for the Details view
